requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.1.0
drf-spectacular==0.26.5
python-dotenv==1.0.0
django-filter==23.5
